rename files in camel case
